estimateWoodProd updates including adding recruitment
At a high level I have included recruitment logic and included that in the woodANPP_Mghayr, expanded the test dataset to include instances of mortality and confirmed that they are producing the desired results, and simplified the code to focus on approach 2 (stand-level). The current script takes as input trees (both single and multi bole trees) from exactly two years, and assumes that the plotSubset is towerAnnualSubset, and for simplicity in v1 I propose that we make these required criteria. Even with just trees, the recruitment logic becomes more complicated when there are 3 or more years worth of data, and as we had previously discussed we would need to have 6 or more years of data for plots sampled only every 5 years. Below I have annotated some of the items in your Git checklist. For the recruitment component you had suggested simply estimating the average biomass of an individual at the minimum diameter threshold. I've done that, at least for v1, applying Chojnacky allometry ID H7 that is essentially a catch-all "other". There may be a better generic allometric equation that could be applied to a generic stem of diameter 10 cm. Alternatively, taxa-specific equations could be applied but that would involve a lot of duplication of effort from the estimateWoodMass function. Since we are talking about small (10 cm diameter) individuals this might not matter.

Line 226-252: The logic and outputs from these lines seem inconsistent with Clark's "Approach 2" as I understand it. In the intermediate 'vst_increment' dataframe, increment is calculated at the scale of the taxonID, which results in a number of NAs for increment when a taxonID is not present in one of the two years. If we assume a stand is roughly equivalent to a plot, Clark Fig 2 indicates to me that all the mass of individuals alive at t2 is summed, and then all the mass of live individuals at t1 is subtracted, and this difference is the "stand" (i.e., plot) level increment; there is no mention of some masses being excluded from the increment calculation based on taxonID. I am confused... why aren't we just using data from the 'vst_agb_kg' data frame as is done for the mortality part of the calculation?
 - the plot-level biomass from the estimateWoodMass should be appropriate for the stand-level increment here (and the summary steps already made). Normally the taxonIDs would sum nicely and we would end up at the same place. Here there is an artifact of an eventID which according to perplotperyear was sampled, yet had no records in the filtered output df, and was subsequently populated with zeros, but not the nlcdClass or taxonID so those fields were NA. For simplicity I have removed the nlcdClass and taxonID grouping variables

Lines 244-252: In the 'vst_increment_long' dataframe, a number of the rows are NA for "Mgha_increment", and I don't understand how this is consistent with Clark "Approach_2".
  - this is an artifact of the issue discussed above. For simplicity and to avoid this edge artifact both the nlcdClass and taxonID grouping variables have been remove

Lines 281-286: After running these lines, there remain a number of individualIDs that have > 1 row with the same "year" and different values for "plantStatus2". I assume this is due to multi-bole 'smt' with different 'plantStatus' for each bole output from 'estimateWoodMass()' function, correct?
 - I am not able to reproduce this (with growthForm = "tree"), so I suspect you are right

Lines 300-306: Using the test dataset, line 301 doesn't seem to do anything to the commas I see in either of the "status_" columns. As such, lines 302-306 also do not seem to do anything. However, if the code worked, I think the assumption being made is if any bole of a multi-bole 'smt' is alive, the entire individual is considered alive, correct? Since individual boles are not tracked, this seems like a reasonable assumption to make. However, it appears that the logic in these lines is hard-coded based on the order of individual boles' status values seen in the data --> what happens when something like '("Live", "Dead_or_Lost", "Live")' is encountered? Or an 'smt' with > 3 boles? Maybe it would be better to use a 'grepl' call and look for "dead"? Regardless, I think this logic isn't working and may be prone to breaking.
 - Thanks for catching this. This must have worked originally, but as you note it was not currently working with the current test dataset. This is now corrected.

 Lines 309-320: Using the test dataset, every row in the new "transitionType_" column is NA, even for individuals like 'NEON.PLA.D02.BLAN.05237' that should have a value of "mortality" if I am understanding the logic. Moreover, a very large number of individualIDs are "Dead_or_Lost" in 2022 and "NULL" in 2021 --> one way this can arise is if an individual is really dead in 2022 but wasn't sampled in 2021 (but was in 2019 - e.g., 'NEON.PLA.D03.JERC.02229'). There are also a very large number of individuals that are "Live" in 2021 and "NULL" in 2022. As it happens, using BLAN_032 as an example, the nested sub area changed from 2021 to 2022 from 25m2 to 10m2, so a number of individuals had their tags removed and were set to plantStatus = "no longer qualifies". Aside from the obvious problems of the code not working as intended, given the complexity of our design, how can we really be confident about estimating mortality for these smaller individuals? I don't know that we can legitimately treat all growthForms the same when it comes to estimating mortality...
 - individuals that experienced mortality (transition from Live to Dead) are now again identified as such. Code now simplified to demand only growthForm of "tree" (single or multi bole tree).

Lines 331-342: Cannot determine whether code is generating expected outcomes because no rows in 'transitions' dataframe have "mortality" in 'transitionsType_' column.
  - upstream issue addressed and a plot with a tree experiencing mortality is now included, so non-zero mortality now produced when expected

 Line 344-351: Cannot determine whether "0" for all rows in "Mgha_mortality" column are legit due to upstream issues described above, but this result does not seem reasonable to me. Also, is "year" incremented here because mortality mass in a given year corresponds to mass in the mass in the previously sampled year, but we want to associate this mass with the productivity for the given year?
  - as for last comment, upstream issue addressed and a plot with a tree experiencing mortality is now included, so non-zero mortality now produced when expected. And yes to the question.

Line 386-393: Hard to tell what this logic is doing because outcome with test dataset does not show any changes before/after running these lines.
  - as for previous 2 comments, upstream issue addressed

Simplified version of script now has code for individual-level approach 1 removed, so last 8 comments not addressed

+++ tests/testthat/test_estimateMass.R
@@ -31,7 +31,7 @@ testthat::test_that(desc = "Output type", {
 testthat::test_that(desc = "Output class biomass_site", {
   testthat::expect_s3_class(object = estimateMassOutputs$biomass_site,
                             class = "data.frame")
-})  
+})
 
 
 
@@ -46,7 +46,7 @@ testthat::test_that(desc = "Output data frame column number", {
 #   Check expected row number of data frame
 testthat::test_that(desc = "Output data frame row number", {
   testthat::expect_identical(object = nrow(estimateMassOutputs$biomass_site),
-                             expected = as.integer(13))
+                             expected = as.integer(15))
 })
 
 
@@ -54,7 +54,7 @@ testthat::test_that(desc = "Output data frame row number", {
 #   Test 'dataProducts' includes appropriate dataProduct option(s)
 testthat::test_that(desc = "Argument 'dataProducts' includes appropriate dataProduct option(s)", {
   testthat::expect_error(object = estimateMass(dataProducts = c("Wat","Air"), # test whether function stops if supplied with an incorrect dataProduct code
-                                               inputDataListVst = VstDat), 
+                                               inputDataListVst = VstDat),
                          regexp = "The dataProducts argument must be one of: 'Bbc', 'Hbp', 'Vst', or a comma-separated combination of two or more of these.")
 })
